cleanup files

+++ env_setup/remove_containers.sh
@@ -6,7 +6,7 @@ remove_containers()
 {
 echo ">> Remvoving the containers..." | tee -a $LOG_FILE
 
-if docker ps | grep "$APACHE_1" >>$LOG_FILE 2>&1
+if docker ps -a | grep "$APACHE_1" >>$LOG_FILE 2>&1
 then
   if docker rm -f $APACHE_1 >>$LOG_FILE 2>&1
   then echo ">> Container $APACHE_1 removed successfully." | tee -a $LOG_FILE
@@ -18,7 +18,7 @@ else echo ">> Container $APACHE_1 not present." | tee -a $LOG_FILE
 fi
 
 
-if docker ps | grep "$APACHE_2" >>$LOG_FILE 2>&1
+if docker ps -a | grep "$APACHE_2" >>$LOG_FILE 2>&1
 then
   if docker rm -f $APACHE_2 >>$LOG_FILE 2>&1
   then echo ">> Container $APACHE_2 removed successfully." | tee -a $LOG_FILE
@@ -30,7 +30,7 @@ else echo ">> Container $APACHE_2 not present." | tee -a $LOG_FILE
 fi
 
 
-if docker ps | grep "$NGINX" >>$LOG_FILE 2>&1
+if docker ps -a | grep "$NGINX" >>$LOG_FILE 2>&1
 then
   if docker rm -f $NGINX >>$LOG_FILE 2>&1
   then echo ">> Container $NGINX removed successfully." | tee -a $LOG_FILE
@@ -42,7 +42,7 @@ else echo ">> Container $NGINX not present." | tee -a $LOG_FILE
 fi
 
 
-if docker ps | grep "$CADVISOR" >>$LOG_FILE 2>&1
+if docker ps -a | grep "$CADVISOR" >>$LOG_FILE 2>&1
 then
   if docker rm -f $CADVISOR >>$LOG_FILE 2>&1
   then echo ">> Container $CADVISOR removed successfully." | tee -a $LOG_FILE
